// src/enums/preparations/preparationAction.ts
export enum PREPARATION_ACTION {
  CUSTOM_PARAMETERS = 'CUSTOM_PARAMETERS',
  DETAIL = 'DETAIL',
  REPEAT = 'REPEAT',
  EDIT = 'EDIT',
  DELETE = 'DELETE',
  ARCHIVE = 'ARCHIVE',
  PHOTO_GALLERY = 'PHOTO_GALLERY',
  CONNECT_DEVICE = 'CONNECT_DEVICE',
  EDIT_DEVICE_CONNECTION = 'EDIT_DEVICE_CONNECTION',
}